Split up prototype code to separate files
