+++ Reqnroll/Bindings/StepDefinitionRegistryAttribute.cs
@@ -3,7 +3,7 @@
 namespace Reqnroll.Bindings;
 
 [AttributeUsage(AttributeTargets.Assembly)]
-public class StepRegistryAttribute(Type registryType) : Attribute
+public class StepDefinitionRegistryAttribute(Type registryType) : Attribute
 {
     public Type RegistryType { get; } = registryType;
 }
